Split background task and TWELITE watcher
