+++ SerialMonitor/MonitorTask.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Nito.AsyncEx;
+using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Windows.ApplicationModel.Background;
@@ -12,41 +13,22 @@ namespace SerialMonitor
 {
     public sealed class MonitorTask : IBackgroundTask
     {
-        private static DeviceWatcher watcher = null;
-        private static bool isEnumerated = false;
-
-        private ISerialCommService commService;
-        private bool IsConnected { get; set; }
-        private SemaphoreSlim semaphore;
         private LogWriter writer;
-        private bool Canceled;
+        private TweLiteWatcher twatcher;
 
         public async void Run(IBackgroundTaskInstance taskInstance)
         {
             var deferral = taskInstance.GetDeferral();
             taskInstance.Canceled += this.OnCanceled;
-            IsConnected = false;
-            semaphore = new SemaphoreSlim(1, 1);
-            commService = new SerialCommService();
-            commService.Received += this.OnReceivedAsync;
             writer = new LogWriter();
             var opened = await writer.OpenAsync();
             if (opened)
             {
-                SetWatcher();
-                while (true)
-                {
-                    await Task.Delay(100);
-                    if (Canceled)
-                    {
-                        break;
-                    }
-                }
-                commService.Stop();
+                twatcher = new TweLiteWatcher(writer);
+                await twatcher.WatchAsync();
+                twatcher.Dispose();
             }
-            commService.Received -= this.OnReceivedAsync;
-            writer.Close();
-            taskInstance.Canceled -= this.OnCanceled;
+            writer.Dispose();
             deferral.Complete();
         }
 
@@ -55,97 +37,7 @@ namespace SerialMonitor
 #pragma warning disable CS4014 // この呼び出しを待たないため、現在のメソッドの実行は、呼び出しが完了する前に続行します
             writer.WriteAsync(reason.ToString());
 #pragma warning restore CS4014 // この呼び出しを待たないため、現在のメソッドの実行は、呼び出しが完了する前に続行します
-            Canceled = true;
-        }
-
-        void SetWatcher()
-        {
-            watcher = DeviceInformation.CreateWatcher(SerialDevice.GetDeviceSelector());
-            watcher.Added += OnAddedAsync;
-            watcher.Removed += OnRemovedAsync;
-            watcher.Updated += OnUpdatedAsync;
-            watcher.EnumerationCompleted += OnEnumerationCompletedAsync;
-            watcher.Start();
-        }
-
-        async void OnAddedAsync(DeviceWatcher sender, DeviceInformation di)
-        {
-            if (isEnumerated && !IsConnected)
-            {
-                await ConnectAsync();
-            }
-        }
-
-        private async Task ConnectAsync()
-        {
-            await semaphore.WaitAsync();
-            try
-            {
-                if (!IsConnected)
-                {
-                    var result = await commService.OpenAsync();
-                    IsConnected = result;
-                    if (result)
-                    {
-#pragma warning disable CS4014 // この呼び出しを待たないため、現在のメソッドの実行は、呼び出しが完了する前に続行します
-                        commService.StartAsync();
-#pragma warning restore CS4014 // この呼び出しを待たないため、現在のメソッドの実行は、呼び出しが完了する前に続行します
-                        await writer.WriteAsync(string.Format("Connected: {0}", commService.Description));
-                    }
-                }
-            }
-            finally
-            {
-                semaphore.Release();
-            }
-        }
-
-        private async void OnReceivedAsync(ICommService sender, MessageEventArgs args)
-        {
-            await writer.WriteAsync(args.Message);
-        }
-
-        private async Task DisconnectAsync()
-        {
-            await semaphore.WaitAsync();
-            try
-            {
-                if (IsConnected)
-                {
-                    commService.Stop();
-                    IsConnected = false;
-                    await writer.WriteAsync(string.Format("Disconnected {0}", commService.Description));
-                }
-            }
-            finally
-            {
-                semaphore.Release();
-            }
-        }
-
-        async void OnUpdatedAsync(DeviceWatcher sender, DeviceInformationUpdate diu)
-        {
-            if (!IsConnected)
-            {
-                await ConnectAsync();
-            }
-        }
-
-        async void OnRemovedAsync(DeviceWatcher sender, DeviceInformationUpdate di)
-        {
-            if (IsConnected && commService.DeviceInfo != null)
-            {
-                if (commService.DeviceInfo.Id.Equals(di.Id))
-                {
-                    await DisconnectAsync();
-                }
-            }
-        }
-
-        private async void OnEnumerationCompletedAsync(DeviceWatcher sender, object args)
-        {
-            isEnumerated = true;
-            await ConnectAsync();
+            twatcher?.Stop();
         }
     }
 }
