+++ SerialMonitor/SerialMonitor.csproj
@@ -113,10 +113,14 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SerialCommService.cs" />
     <Compile Include="MonitorTask.cs" />
+    <Compile Include="TweLiteWatcher.cs" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.NETCore.UniversalWindowsPlatform">
       <Version>6.0.8</Version>
+    </PackageReference>
+    <PackageReference Include="Nito.AsyncEx">
+      <Version>4.0.1</Version>
     </PackageReference>
   </ItemGroup>
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '14.0' ">
